Use combine's partial parsing support
`combine` has (unstable, thus the git dependency) support for partial
parsing which lets parsers constructed using it recover from incomplete
inputs and later resume parsing at the point that it stopped (or close
to it).

Though the implementation details are still volatile it appears to work
very well and this served as a good test case for it.

Naturally combine will need a proper release with this feature before
this change is even possible to include in this project.